requirements.txt
certifi==2017.11.5
chardet==3.0.4
feedparser==5.2.1
future==0.16.0
idna==2.6
python-telegram-bot==9.0.0
requests==2.18.4
schedule==0.5.0
urllib3==1.22
